curl: improve run-ptest
There's no need to run the output of runtests.pl through a sed to get
automake-style output, as you can pass -am to get this formatting.

Don't run timing dependent tests, as the ptests can run on loaded
systems.

Add a dependency on the en_US locale because some of the tests require
this.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>